server root_path change and refactoring
