+++ .gitignore
@@ -1,5 +1,14 @@
 test.ipynb
+sqls.sql
 app/__pycache__
+app/auth/__pycache__
+app/auth/__pycache__
+app/database/__pycache__
+app/model/__pycache__
+app/routes/__pycache__
+app/routes/api/__pycache__
+app/routes/apihelper/__pycache__
+app/.env
 
 
 # gitignore가 작동하지 않을때 대처법
